General fixes and improvements
Fixed bug when exporting Measurements; added example files for SLA and
timeline;

+++ core/src/main/java/br/ufc/lsbd/benchxtend/LogEntry.java
@@ -9,10 +9,10 @@ package br.ufc.lsbd.benchxtend;
  * @author rodrigofelix
  */
 public class LogEntry {
-    public int time;
+    public long time;
     public int numberClients;
     
-    public LogEntry(int time, int numberClients){
+    public LogEntry(long time, int numberClients){
         this.time = time;
         this.numberClients = numberClients;
     }
